+++ packages/react/src/components/PlanDisplay/src/components/PlanDisplayCardPrice/index.ts
@@ -1,0 +1,5 @@
+import { PlanDisplayCardPrice } from "./PlanDisplayCardPrice";
+
+export { PlanDisplayCardPrice };
+
+export default PlanDisplayCardPrice;